Refactor storage management and update ModEntry
- Changed `ModEntry` access level to `internal sealed` and added new data structures for storage management.
- Introduced `TryGetDataDelegate` for data retrieval, replacing the `StorageManager` reference in `ModPatches`.
- Removed the `StorageManager` class and its associated logic.
- Cleared contents of `default.json` configuration file.
