--- ExpandedStorage/i18n/default.json
@@ -1,4 +1,0 @@
-{
-  "config.section.general.title": "General",
-  "config.section.general.description": "General settings for this mod."
-}
